OneDrive/Desktop/trace-memory-copilot-highlevel/trace-memory-copilot-highlevel/pyproject.toml
[project]
name = "trace-memory-copilot"
version = "1.0.0"
description = "TraceMemory Copilot: a Zep-powered assistant with long-term memory and a Streamlit UI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ag2[ollama]>=0.9",
  "ollama>=0.4.8",
  "streamlit>=1.44.1",
  "zep-cloud>=2.11.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
package = false
